BISERVER-1683
Spoke with the MSC team (and verified) that they've removed the optimizations discussed. If they do use the optimizations, they keep a local copy of the gwt.xml with the optimizations and don't check it in.  I added a comment to the gwt.xml files stating that optimizations are not to be put in production files.

git-svn-id: svn://source.pentaho.org/svnroot/pentaho-open-admin-console/trunk@32441 5d404267-db13-0410-8476-b1f6b398e63a